requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
tqdm>=4.62.0
requests>=2.26.0

# Web interface
fastapi>=0.68.0
uvicorn>=0.15.0
gradio>=3.0.0
python-multipart>=0.0.5

# Image processing
Pillow>=8.3.0
opencv-python>=4.5.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Audio processing
librosa>=0.8.0

# Machine learning and NLP
scikit-learn>=0.24.0
transformers>=4.10.0

# Logging and utilities
python-json-logger>=2.0.0

# Optional: for development
pytest>=6.2.0
black>=21.7b0
flake8>=3.9.0
mypy>=0.910 